net: move generic connection to net::
